backend/pyproject.toml
# pyproject.toml
# ─────────────────────────────────────────────────────────────────
# LEARNING NOTE – pyproject.toml
#
# pyproject.toml is the modern Python project config file (PEP 517/518/621).
# It replaces setup.py, setup.cfg, requirements.txt and .flake8 in one file.
#
# Sections:
#   [build-system]  – which build backend to use (hatchling here, very simple)
#   [project]       – package metadata + runtime dependencies
#   [project.optional-dependencies] – extras (test, dev tools)
#   [tool.ruff]     – linter/formatter config (CP4)
#   [tool.pytest…]  – pytest config (CP4)
# ─────────────────────────────────────────────────────────────────

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "myteams-backend"
version = "0.1.0"
description = "MyTeams FastAPI backend – learning project"
requires-python = ">=3.12"

dependencies = [
    # FastAPI – the web framework
    "fastapi>=0.111.0",
    # Uvicorn – the ASGI server that runs FastAPI
    "uvicorn[standard]>=0.29.0",
    # Pydantic Settings – reads config from env vars / .env files
    "pydantic-settings>=2.2.0",
    # SQLAlchemy – ORM + query builder (CP6)
    # Lets us write Python classes instead of raw SQL.
    # Swapping SQLite → Postgres later only requires changing the URL.
    "sqlalchemy>=2.0.0",
    # psycopg2 – Postgres driver (CP7)
    # The adapter that lets SQLAlchemy talk to Postgres.
    # `binary` variant includes precompiled C extensions – no libpq needed.
    "psycopg2-binary>=2.9.0",
    # python-jose – JWT creation and verification (CP9)
    "python-jose[cryptography]>=3.3.0",
    # passlib – password hashing (CP9)
    # bcrypt extra pulls in the bcrypt C extension for fast hashing.
    # Pin bcrypt<4 – passlib 1.7.x doesn't support the bcrypt 4+ API changes.
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    # email-validator – required by Pydantic EmailStr (CP9)
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    # ruff – extremely fast linter + formatter (replaces flake8, black, isort)
    "ruff>=0.4.0",
    # pytest – test runner
    "pytest>=8.0.0",
    # httpx – async-capable HTTP client; needed for FastAPI TestClient
    "httpx>=0.27.0",
    # mypy – static type checker (CP4)
    # Checks that types actually match at every call site, which ruff/pyflakes
    # cannot do.  Stubs packages (mypy-extensions, types-*) teach mypy about
    # third-party libraries that ship without inline type hints.
    "mypy>=1.10.0",
]

# ─────────────────────────────────────────────────────────────────
# Ruff configuration (used in CP4 Quality Gates)
# We configure it now so the file is ready when we wire CI.
# ─────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# E/W = pycodestyle, F = pyflakes, I = isort, UP = pyupgrade
select = ["E", "W", "F", "I", "UP"]
# UP017: suggests datetime.UTC (module attr) but we import the datetime class
# directly, so timezone.utc is the correct way to reference it.
ignore = ["UP017"]

[tool.ruff.lint.per-file-ignores]
# Test files often have intentionally long one-liners for clarity.
"app/tests/*.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ─────────────────────────────────────────────────────────────────
# Pytest configuration
# ─────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

# ─────────────────────────────────────────────────────────────────
# mypy configuration (CP4 – Static type checking)
#
# LEARNING NOTE – what mypy does that ruff cannot:
#   ruff = style + obvious mistakes (unused imports, wrong syntax)
#   mypy = type correctness (passing a str where int expected, None safety…)
#
# strict = True turns on all checks.  We ignore missing stubs for
# third-party libs (fastapi, pydantic) that ship their own types.
# ─────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
# FastAPI and Pydantic ship their own type info – no stubs needed
ignore_missing_imports = true
# Don't type-check test files strictly (they use fixtures, no annotations)
exclude = ["app/tests/"]
